+++ projects/angular2-draggable/src/lib/angular-draggable.module.ts
@@ -1,14 +1,17 @@
 import { NgModule } from '@angular/core';
 import { AngularDraggableDirective } from './angular-draggable.directive';
+import { AngularResizableDirective } from './angular-resizable.directive';
 
 @NgModule({
   imports: [
   ],
   declarations: [
-    AngularDraggableDirective
+    AngularDraggableDirective,
+    AngularResizableDirective
   ],
   exports: [
-    AngularDraggableDirective
+    AngularDraggableDirective,
+    AngularResizableDirective
   ]
 })
 export class AngularDraggableModule { }