+++ projects/angular2-draggable/src/lib/models/resize-handle-type.ts
@@ -1,0 +1,12 @@
+export interface ResizeHandleStyle {
+  'n'?: string;
+  's'?: string;
+  'e'?: string;
+  'w'?: string;
+  'ne'?: string;
+  'nw'?: string;
+  'se'?: string;
+  'sw'?: string;
+}
+
+export type ResizeHandleType = string|ResizeHandleStyle;