+++ projects/angular2-draggable/src/public_api.ts
@@ -3,4 +3,6 @@
  */
 
 export * from './lib/angular-draggable.directive';
+export * from './lib/angular-resizable.directive';
 export * from './lib/angular-draggable.module';
+export * from './lib/models/position';