+++ src/app/views/home/home.component.ts
@@ -8,7 +8,18 @@ import { Component, OnInit } from '@angular/core';
 export class HomeComponent implements OnInit {
   demo_html = `\
 <!-- Basic Usage -->
-<div ngDraggable>Drag Me!</div>
+<div ngDraggable> Drag Me! </div>
+
+<!-- Disable -->
+<div [ngDraggable]="false"> Disabled </div>
+`;
+
+  demo_rz_html = `\
+<!-- Basic Usage -->
+<div ngResizable> I'm resizable </div>
+
+<!-- Disable -->
+<div [ngResizable]="false"> Disabled </div>
 `;
 
   constructor() { }